feat: add support for jsx v4
With rescript v10.1 the compiler switched to the newest version of jsx (v4). Both versions are now
supported and can be selected via plugin options. V4 is the new default

BREAKING CHANGE: JSX v4 is the new default, to use this plugin wiht older versions of rescript,
`{"jsx": "v3"}` has to be set via plugin options.