fix: add reactdom to possible component names
The v4 compiler in classic mode can create components with "ReactDOM.createElement". In v3 it only
used "ReactDOMRe". Also switched to v3 as default compiler to be consistent with the rescript
compiler behaviour

BREAKING CHANGE: JSX v3 is now the default. Option `{jsx: "v4"}` has to be used for v4 in automatic
mode.